Resolved different bugs while parsing json data, now response successfully parsed.

+++ app/src/main/java/com/softaai/upstoxholding/data/model/Data.kt
@@ -9,7 +9,7 @@ data class Data(
     @Json(name = "avg_price")
     val avgPrice: String,
     @Json(name = "close")
-    val close: Int,
+    val close: Float,
     @Json(name = "cnc_used_quantity")
     val cncUsedQuantity: Int,
     @Json(name = "collateral_qty")
